+++ app/assets/javascripts/discourse/app/components/modal/login.hbs
@@ -4,7 +4,7 @@
   @closeModal={{@closeModal}}
   @flash={{this.flash}}
   @flashType={{this.flashType}}
-  {{did-insert this.doTheThing}}
+  {{did-insert this.preloadLogin}}
 >
   <:body>
     <PluginOutlet @name="login-before-modal-body" @connectorTagName="div" />
